storage crate builds under default features
