+++ raphtory-memstorage/src/db/api/mod.rs
@@ -1,1 +1,2 @@
 pub mod storage;
+pub mod list_ops;
