+++ raphtory-memstorage/src/db/api/storage/graph/edges/mod.rs
@@ -1,1 +1,5 @@
 pub mod mem_edge;
+pub mod edge_entry;
+pub mod edge_ref;
+pub mod edges;
+pub mod unlocked;
