+++ raphtory-memstorage/src/db/api/storage/graph/nodes/node_owned_entry.rs
@@ -3,6 +3,7 @@ use crate::disk_graph::storage_interface::node::DiskOwnedNode;
 
 #[cfg(feature = "storage")]
 use either::Either;
+use raphtory_api::core::{entities::{edges::edge_ref::EdgeRef, LayerIds}, Direction};
 
 use crate::core::{
         entities::nodes::node_store::NodeStore, storage::ArcEntry
@@ -33,8 +34,8 @@ macro_rules! for_all_iter {
     }};
 }
 
-// impl NodeStorageIntoOps for NodeOwnedEntry {
-//     fn into_edges_iter(self, layers: LayerIds, dir: Direction) -> impl Iterator<Item = EdgeRef> {
-//         for_all_iter!(self, node => node.into_edges_iter(layers, dir))
-//     }
-// }
+impl NodeOwnedEntry {
+    pub fn into_edges_iter(self, layers: &LayerIds, dir: Direction) -> impl Iterator<Item = EdgeRef> {
+        for_all_iter!(self, node => node.into_edges(layers, dir))
+    }
+}
