+++ raphtory-memstorage/src/lib.rs
@@ -1,2 +1,10 @@
 pub mod core;
 pub mod db;
+
+pub enum FilterState {
+    Neither,
+    Both,
+    BothIndependent,
+    Nodes,
+    Edges,
+}
