+++ raphtory/src/db/api/state/node_state.rs
@@ -11,72 +11,6 @@ use std::{
     borrow::Borrow, collections::HashMap, fmt::Debug, hash::Hash, marker::PhantomData, sync::Arc,
 };
 
-#[derive(Clone, Debug)]
-pub struct Index<K> {
-    keys: Arc<[K]>,
-    map: Arc<HashMap<K, usize>>,
-}
-
-impl<K: Copy + Hash + Eq> From<Vec<K>> for Index<K> {
-    fn from(keys: Vec<K>) -> Self {
-        let map = keys
-            .iter()
-            .copied()
-            .enumerate()
-            .map(|(i, k)| (k, i))
-            .collect();
-        Self {
-            keys: keys.into(),
-            map: Arc::new(map),
-        }
-    }
-}
-
-impl<K: Copy + Hash + Eq + Send + Sync> Index<K> {
-    pub fn iter(&self) -> impl Iterator<Item = &K> + '_ {
-        self.keys.iter()
-    }
-
-    pub fn into_par_iter(self) -> impl IndexedParallelIterator<Item = K> {
-        let keys = self.keys;
-        (0..keys.len()).into_par_iter().map(move |i| keys[i])
-    }
-
-    pub fn into_iter(self) -> impl Iterator<Item = K> {
-        let keys = self.keys;
-        (0..keys.len()).map(move |i| keys[i])
-    }
-
-    pub fn index<Q: ?Sized>(&self, key: &Q) -> Option<usize>
-    where
-        K: Borrow<Q>,
-        Q: Hash + Eq,
-    {
-        self.map.get(key).copied()
-    }
-
-    pub fn key(&self, index: usize) -> Option<K> {
-        self.keys.get(index).copied()
-    }
-
-    pub fn len(&self) -> usize {
-        self.keys.len()
-    }
-
-    pub fn contains<Q: ?Sized>(&self, key: &Q) -> bool
-    where
-        K: Borrow<Q>,
-        Q: Hash + Eq,
-    {
-        self.map.contains_key(key)
-    }
-}
-
-impl<K: Copy + Hash + Eq + Send + Sync> Index<K> {
-    pub fn par_iter(&self) -> impl IndexedParallelIterator<Item = &K> + '_ {
-        self.keys.par_iter()
-    }
-}
 
 pub struct NodeState<'graph, V, G, GH = G> {
     base_graph: G,
